Move dataset to own file
